Rename ray_sampler_utils to scripted policy
- Remove mock_algo in favor of Mocked object algo
- generalize scripted policy for more than 4x4 gridworld
