+++ tests/fixtures/policies/__init__.py
@@ -1,6 +1,6 @@
 from tests.fixtures.policies.dummy_policy import DummyPolicy
 from tests.fixtures.policies.dummy_recurrent_policy import (
     DummyRecurrentPolicy)
-from tests.fixtures.policies.ray_sampler_utils import ScriptedPolicy
+from tests.fixtures.policies.scripted_policy import ScriptedPolicy
 
 __all__ = ['DummyPolicy', 'DummyRecurrentPolicy', 'ScriptedPolicy']
