+++ tests/garage/sampler/test_ray_sampler.py
@@ -8,21 +8,42 @@ from garage.sampler import RaySampler, SamplerWorker
 from garage.tf.envs import TfEnv
 from garage.tf.samplers.on_policy_vectorized_sampler \
     import OnPolicyVectorizedSampler
-from tests.fixtures.policies.ray_sampler_utils import ScriptedPolicy
+from tests.fixtures.policies.scripted_policy import ScriptedPolicy
 
 
 class TestSampler():
+    """
+    Uses mock policy for 4x4 gridworldenv
+    '4x4': [
+        'SFFF',
+        'FHFH',
+        'FFFH',
+        'HFFG'
+    ]
+    0: left
+    1: down
+    2: right
+    3: up
+    -1: no move
+    'S' : starting point
+    'F' or '.': free space
+    'W' or 'x': wall
+    'H' or 'o': hole (terminates episode)
+    'G' : goal
+    [2,2,1,0,3,1,1,1,2,2,1,1,1,2,2,1]
+    """
+
     def setup_method(self):
         ray.init(local_mode=True, ignore_reinit_error=True)
 
         self.env = TfEnv(GridWorldEnv(desc='4x4'))
-        self.policy = ScriptedPolicy()
+        self.policy = ScriptedPolicy(
+            scripted_actions=[2, 2, 1, 0, 3, 1, 1, 1, 2, 2, 1, 1, 1, 2, 2, 1])
         self.algo = Mock(
             env_spec=self.env.spec, policy=self.policy, max_path_length=16)
 
     def teardown_method(self):
         self.env.close()
-        ray.shutdown()
 
     def test_ray_batch_sampler(self):
         sampler1 = RaySampler(
@@ -57,3 +78,5 @@ class TestSampler():
             assert (np.array_equal(trajectory['observations'], true_obs))
             assert (np.array_equal(trajectory['actions'], true_actions))
             assert (np.array_equal(trajectory['rewards'], true_rewards))
+        sampler1.shutdown_worker()
+        sampler2.shutdown_worker()
